Pushing incomplete v 1.0.2
- Moving game server code into unity, so can have pickups/enemies
- optimising packets being sent/received

+++ Bioforce-2D/Assets/Entities/Player/Scripts/Gun/IGun.cs
@@ -3,7 +3,7 @@
 public interface IGun
 {
     void ShootBullet(Vector2 position, Quaternion rotation);
-    void SetOwnerClientID(int iD);
+    void SetOwnerClientID(byte iD);
     void Disable(TypeOfDeath typeOfDeath);
     void Enable();
     void SetColor(Color playerColor);
