+++ Bioforce-2D/Assets/Networking/InternetDiscoveryClient.cs
@@ -3,12 +3,12 @@ public class InternetDiscoveryClient : DiscoveryTCPClient
 {
     protected override void InitPacketHandlerDictionary()
     {
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.welcome, ServerMenu.ReadWelcomePacket);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.serverData, ServerMenu.ReadServerDataPacket);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.serverDeleted, ServerMenu.ReadServerDeletedPacket);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.serverModified, ServerMenu.ReadServerModifiedPacket);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.cantJoinServerDeleted, ServerMenu.ReadCantJoinServerDeleted);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.noMoreServersAvailable, ServerMenu.ReadNoMoreServersAvailable);
-        PacketHandlerDictionary.Add((int)InternetDiscoveryServerPackets.joinServer, ServerMenu.ReadJoinServer);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.welcome, ServerMenu.ReadWelcomePacket);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.serverData, ServerMenu.ReadServerDataPacket);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.serverDeleted, ServerMenu.ReadServerDeletedPacket);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.serverModified, ServerMenu.ReadServerModifiedPacket);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.cantJoinServerDeleted, ServerMenu.ReadCantJoinServerDeleted);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.noMoreServersAvailable, ServerMenu.ReadNoMoreServersAvailable);
+        PacketHandlerDictionary.Add((byte)InternetDiscoveryServerPackets.joinServer, ServerMenu.ReadJoinServer);
     }
 }
